Fixed issues with FRAME_COUNT. Now it works and stores count+1 instead of count*2.
